fix: 기존 upStationId와 downStationId 필드의 데이터 타입을 Long에서 int으로 변경
- refactor: 지하철 노선 생성 테스트 메서드 리팩토링

+++ src/main/java/subway/dto/StationLineRequest.java
@@ -3,13 +3,13 @@ package subway.dto;
 public class StationLineRequest {
     private final String name;
     private final String color;
-    private final Long upStationId;
+    private final int upStationId;
 
-    private final Long downStationId;
+    private final int downStationId;
 
-    private final Long distance;
+    private final int distance;
 
-    public StationLineRequest(String name, String color, Long upStationId, Long downStationId, Long distance) {
+    public StationLineRequest(String name, String color, int upStationId, int downStationId, int distance) {
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
@@ -25,15 +25,15 @@ public class StationLineRequest {
         return color;
     }
 
-    public Long getUpStationId() {
+    public int getUpStationId() {
         return upStationId;
     }
 
-    public Long getDownStationId() {
+    public int getDownStationId() {
         return downStationId;
     }
 
-    public Long getDistance() {
+    public int getDistance() {
         return distance;
     }
 }
